+++ src/components/series/Series.tsx
@@ -221,6 +221,12 @@ const Series: FC<RouteComponentProps<{ id: string }>> = () => {
   // 保存
   const handleSave = () => {
     // console.log('handleSave');
+    let role = localStorage.getItem('fragile')
+    role = role ? window.atob(role) : '3'
+    if (`${role}` === '3') {
+      message.error(`访客模式，不能${aesMode === 1 ? '提交' : '保存'}`)
+      return
+    }
     form
       .validateFields()
       .then(async (values: SeriesT) => {
@@ -281,6 +287,12 @@ const Series: FC<RouteComponentProps<{ id: string }>> = () => {
   // 删除
   const handleDelete = async (id: React.Key) => {
     // console.log('handleDelete', id);
+    let role = localStorage.getItem('fragile')
+    role = role ? window.atob(role) : '3'
+    if (`${role}` === '3') {
+      message.error('访客模式，不能删除')
+      return
+    }
     try {
       const res = await deleteSeries({ id })
       if (res?.error_code === '00') {
